requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0

# Development and experimentation
jupyter>=1.0.0
matplotlib>=3.7.0
tensorboard>=2.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
